REORDER-AT-REBASE: ofi: prevent misuse of GASNETC_OFI_HAS_MR_SCALABLE
AT REBASE - make this commit come before the "real work".

This commit ensures that `GASNETC_OFI_HAS_MR_SCALABLE` cannot be
(mis)used in preprocessor directives.  By adding an explicit cast to
the definition, the value as a C expression remains unchanged, but use
in a preprocessor expression becomes a syntax error.

+++ ofi-conduit/gasnet_ofi.c
@@ -95,7 +95,8 @@ static uint64_t* gasnetc_ofi_target_aux_keys;
 #ifdef GASNETC_OFI_HAS_MR_SCALABLE
   #define GASNETC_OFI_HAS_MR_SCALABLE_STATIC 1
 #else
-  #define GASNETC_OFI_HAS_MR_SCALABLE has_mr_scalable
+  // cast prevents erroneous use in preprocessor directives
+  #define GASNETC_OFI_HAS_MR_SCALABLE ((short)has_mr_scalable)
 #endif
 
 // Alias unless/until the properties are split
